Updated Report Draft
Report draft now contains expectations and weightings for various
sections.

Sidenote: most of the report should be written with influence from both
members of the group, attention given where sections are lacking. I
don’t think it’s important or required that we attempt to divvy up the
workload.

The report is worth 40% of project mark, a fair chunk.